requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision
torchaudio
matplotlib==3.10.3
numpy==2.2.6
opencv_python==4.11.0.86
Pillow==11.2.1
torchmetrics==1.7.3
torch-fidelity==0.3.0
tqdm==4.67.1
